requirements.txt
# Core video processing
yt-dlp>=2023.11.16
ffmpeg-python>=0.2.0

# Web interface
streamlit>=1.24.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
